Renamed a package from exeptions to exceptions
