+++ src/com/pentalog/pentastagiu/homework6/tests/ATMTests.java
@@ -1,8 +1,8 @@
 package com.pentalog.pentastagiu.homework6.tests;
 
 import com.pentalog.pentastagiu.homework6.business.CashDispenser;
-import com.pentalog.pentastagiu.homework6.exeptions.ClientNotFoundException;
-import com.pentalog.pentastagiu.homework6.exeptions.InsufficientFundsException;
+import com.pentalog.pentastagiu.homework6.exceptions.ClientNotFoundException;
+import com.pentalog.pentastagiu.homework6.exceptions.InsufficientFundsException;
 import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
